add functional.py
