+++ farmer/ncc/encoder_layer/__init__.py
@@ -1,2 +1,3 @@
 from .dilated_xception import *
 from .mobilenetv2 import *
+from .functional import *
